ref: clean up unused features
