+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,10 +1,10 @@
 ---
 name: ⚒  Feature request
 about: Suggest an idea for this project
-title: ''
-labels: 'feature'
-assignees: ''
-projects: 'webapp'
+title: ""
+labels: "feature"
+assignees: ""
+projects: "webapp"
 ---
 
 ## Descrição
@@ -16,7 +16,7 @@ A clear and concise description of what the problem is. Ex. I'm always frustrate
 
 ## Links uteis
 
-- [Planing](https://podcodar.notion.site/Sprint-Planing-2-203c837559494a9887af633792c095ee)
+- [Planning](https://podcodar.notion.site/Sprint-Planing-2-203c837559494a9887af633792c095ee)
 - [figma template](https://www.notion.so/podcodar/Sprint-Planing-2-203c837559494a9887af633792c095ee#e4c586f975564d0cb8310489432e0c64)
 - [Choc UI](https://www.notion.so/podcodar/Sprint-Planing-2-203c837559494a9887af633792c095ee#3fbfcf4f78d142978b8a29b190f0f17a)
 - [Eng wiki > Links uteis](https://www.notion.so/podcodar/Chakra-1e4429a361a844f9a5d5db6790dd154b#e624ab829d754dc591ee3c6bfd1b1d76)
