+++ README.md
@@ -4,7 +4,7 @@ It's nice to have you here! This project is the web app created by the PodCodar 
 
 ### Technologies
 
-<div  style="max-width: 200px; display: grid; grid-template-columns: repeat(6,auto); grid-gap: 1rem;">
+<div style="max-width: 200px; display: grid; grid-template-columns: repeat(6,auto); grid-gap: 1rem;">
   <img src="./docs/images/typescript.svg" height="30" />
   <img src="./docs/images/react.svg" height="30" />
   <img src="./docs/images/next.svg" height="30" />
@@ -13,29 +13,28 @@ It's nice to have you here! This project is the web app created by the PodCodar 
   <img src="./docs/images/linux.svg" height="30" />
 </div>
 
-### Partners
-
-[![image](./docs/images/vercel-logo.svg)][1]
-
 ### Start Project
 
-- Install dependencies using library yarn whit command `yarn`.
+To work on this project you have use `yarn` or `bun` as package manager.
 
-- run `yarn decrypt` to decrypt the `.env.gpg` file. (ask the password for the current teammembers)
+> Personal suggestion: use `bun` and be happy.
 
-- now enter a command `yarn dev` your project will be running.
+- Install dependencies using library bun whit command `bun`.
 
-- to run de **production** build `yarn build && yarn start`
+- run `bun decrypt` to decrypt the `.env.gpg` file. (ask the password for the current teammembers)
+
+- now enter a command `bun dev` your project will be running.
+
+- to run de **production** build `bun build && bun start`
 
 ### Our Wikis
 
 - [WebApp wiki page][2]
 - [Engineering Wiki][3]
-- [Project Board][4]
+- [BanBan Board][4]
 
 ### Let's code with us
 
-[1]: https://vercel.com/?utm_source=podcodar&utm_campaign=oss
 [2]: https://www.notion.so/podcodar/PodCodar-WebApp-4c1b6a896ebf45b49fed62c28ac45bef
 [3]: https://www.notion.so/podcodar/Engineering-Wiki-597eb2a7a0534547b1e0f9a2e48d22d2
 [4]: https://github.com/orgs/podcodar/projects/4
