+++ packages/hooks/useIsClient.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react";
 
 export function useIsClient() {
-	const [isClient, setIsClient] = useState(false);
+  const [isClient, setIsClient] = useState(false);
 
-	useEffect(() => {
-		setIsClient(true);
-	}, []);
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
 
-	return isClient && typeof window !== "undefined";
+  return isClient && typeof window !== "undefined";
 }
