--- packages/utils/functions.ts
@@ -1,7 +1,0 @@
-export function makeThrowMissingImplementation(message = "missing implementation") {
-  return function throwMissingImplementation() {
-    throw new Error(message);
-  };
-}
-
-export type Result<V, E = Error> = [V, null] | [null, E];
